Seperate ToolTable and Tool implementations
Seperate the tooltable and tool implementations.
